change encoding logic and simplify tests.
